scsi: zfcp: Change the type of all fsf request id fields and variables to u64
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-494.el8
commit-author Benjamin Block <bblock@linux.ibm.com>
commit 3ab01810153b0ca26cfc846ba72996b250adde6d

We use different integer types throughout zfcp to store the FSF request ID
and related values; some places use 'unsigned long' and others 'u64'.  On
s390x these are effectively the same type, but this might cause confusions
and is generally inconsistent.

The specification for the used hardware specifies this value as a 64-bit
number, and ultimately we use this value to communicate with the hardware,
so it makes sense to change the type of all these variables to 'u64' where
we can.  The only exception being when we store it in the 'host_scribble'
field of a 'struct scsi_cmnd'; for this case we add a build time check to
make sure they are compatible.

Link: https://lore.kernel.org/r/9c9cbe5acc2b419a22dce2fed847e3db91b60201.1677000450.git.bblock@linux.ibm.com
	Signed-off-by: Benjamin Block <bblock@linux.ibm.com>
	Reviewed-by: Steffen Maier <maier@linux.ibm.com>
	Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
(cherry picked from commit 3ab01810153b0ca26cfc846ba72996b250adde6d)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>