dm verity: fix error handling for check_at_most_once on FEC
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-494.el8
commit-author Yeongjin Gil <youngjin.gil@samsung.com>
commit e8c5d45f82ce0c238a4817739892fe8897a3dcc3

In verity_end_io(), if bi_status is not BLK_STS_OK, it can be return
directly. But if FEC configured, it is desired to correct the data page
through verity_verify_io. And the return value will be converted to
blk_status and passed to verity_finish_io().

BTW, when a bit is set in v->validated_blocks, verity_verify_io() skips
verification regardless of I/O error for the corresponding bio. In this
case, the I/O error could not be returned properly, and as a result,
there is a problem that abnormal data could be read for the
corresponding block.

To fix this problem, when an I/O error occurs, do not skip verification
even if the bit related is set in v->validated_blocks.

Fixes: 843f38d382b1 ("dm verity: add 'check_at_most_once' option to only validate hashes once")
	Cc: stable@vger.kernel.org
	Reviewed-by: Sungjong Seo <sj1557.seo@samsung.com>
	Signed-off-by: Yeongjin Gil <youngjin.gil@samsung.com>
	Signed-off-by: Mike Snitzer <snitzer@kernel.org>
(cherry picked from commit e8c5d45f82ce0c238a4817739892fe8897a3dcc3)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>