scsi: target: core: Make logs less verbose
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-494.el8
commit-author Konstantin Shelekhin <k.shelekhin@yadro.com>
commit 05787e3456ffcc8598aab632fcd5d160894619b3

Change the log level of the following message to debug:

	Unsupported SCSI Opcode 0xXX, sending CHECK_CONDITION.

This message is mostly helpful during debugging sessions in order to
understand errors on the initiator side. But most of the time it's just
useless and makes reading logs much harder.

It gets particularly annoying if there are many initiators that come and go
or if an initiator runs a program that does not care whether the command is
supported and just keeps sending it.

Link: https://lore.kernel.org/r/20210929114959.705852-1-k.shelekhin@yadro.com
	Reviewed-by: Roman Bolshakov <r.bolshakov@yadro.com>
	Reviewed-by: Lee Duncan <lduncan@suse.com>
	Signed-off-by: Konstantin Shelekhin <k.shelekhin@yadro.com>
	Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
(cherry picked from commit 05787e3456ffcc8598aab632fcd5d160894619b3)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>