kunit: tool: Report an error if any test has no subtests
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-494.el8
commit-author David Gow <davidgow@google.com>
commit e56e482855b782b0e38b91cc383324bb80461073

It's possible for a test to have a subtest header, but zero valid
subtests. We used to error on this if the test plan had no subtests
listed, but it's possible to have subtests without a test plan (indeed,
this is how parameterised tests work).

Tests with 0 subtests now have the result NO_TESTS, and will report an
error (which does not halt test execution, but is printed in a scary red
colour and is noted in the results summary).

	Signed-off-by: David Gow <davidgow@google.com>
	Reviewed-by: Daniel Latypov <dlatypov@google.com>
	Reviewed-by: Brendan Higgins <brendanhiggins@google.com>
	Signed-off-by: Shuah Khan <skhan@linuxfoundation.org>
(cherry picked from commit e56e482855b782b0e38b91cc383324bb80461073)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>