kunit: tool: Default --jobs to number of CPUs
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-494.el8
commit-author David Gow <davidgow@google.com>
commit ad659ccb5412874c6a89d3588cb18857c00e9d0f

The --jobs parameter for kunit_tool currently defaults to 8 CPUs,
regardless of the number available. For systems with significantly more
(or less), this is not as efficient. Instead, default --jobs to the
number of CPUs available to the process: while there are as many
superstitions as to exactly what the ideal jobs:CPU ratio is, this seems
sufficiently sensible to me.

A new helper function to get the default number of jobs is added:
get_default_jobs() -- this is used in kunit_tool_test instead of a
hardcoded value, or an explicit call to len(os.sched_getaffinity()), so
should be more flexible if this needs to change in the future.

	Signed-off-by: David Gow <davidgow@google.com>
	Signed-off-by: Daniel Latypov <dlatypov@google.com>
	Reviewed-by: Daniel Latypov <dlatypov@google.com>
	Reviewed-by: Brendan Higgins <brendanhiggins@google.com>
	Signed-off-by: Shuah Khan <skhan@linuxfoundation.org>
(cherry picked from commit ad659ccb5412874c6a89d3588cb18857c00e9d0f)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>