linux/kernel.h: Use parentheses around argument in u64_to_user_ptr()
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-494.el8
commit-author Jann Horn <jannh@google.com>
commit a0fe2c6479aab5723239b315ef1b552673f434a3

Use parentheses around uses of the argument in u64_to_user_ptr() to
ensure that the cast doesn't apply to part of the argument.

There are existing uses of the macro of the form

  u64_to_user_ptr(A + B)

which expands to

  (void __user *)(uintptr_t)A + B

(the cast applies to the first operand of the addition, the addition
is a pointer addition). This happens to still work as intended, the
semantic difference doesn't cause a difference in behavior.

But I want to use u64_to_user_ptr() with a ternary operator in the
argument, like so:

  u64_to_user_ptr(A ? B : C)

This currently doesn't work as intended.

	Signed-off-by: Jann Horn <jannh@google.com>
	Signed-off-by: Borislav Petkov <bp@suse.de>
	Reviewed-by: Mukesh Ojha <mojha@codeaurora.org>
	Cc: Andrei Vagin <avagin@openvz.org>
	Cc: Andrew Morton <akpm@linux-foundation.org>
	Cc: Dan Carpenter <dan.carpenter@oracle.com>
	Cc: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
	Cc: "H. Peter Anvin" <hpa@zytor.com>
	Cc: Ingo Molnar <mingo@kernel.org>
	Cc: Jani Nikula <jani.nikula@intel.com>
	Cc: Kees Cook <keescook@chromium.org>
	Cc: Masahiro Yamada <yamada.masahiro@socionext.com>
	Cc: NeilBrown <neilb@suse.com>
	Cc: Peter Zijlstra <peterz@infradead.org>
	Cc: Qiaowei Ren <qiaowei.ren@intel.com>
	Cc: Thomas Gleixner <tglx@linutronix.de>
	Cc: x86-ml <x86@kernel.org>
Link: https://lkml.kernel.org/r/20190329214652.258477-1-jannh@google.com
(cherry picked from commit a0fe2c6479aab5723239b315ef1b552673f434a3)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>