tracing/kprobes: Fix within_notrace_func() to check only notrace functions
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-496.el8
commit-author Masami Hiramatsu <mhiramat@kernel.org>
commit 6bc6c77cfc6747292d5865d124cc31333b1b8536

Fix within_notrace_func() to check only notrace functions and to ignore the
kprobe-event which can not solve symbol addresses.

within_notrace_func() returns true if the given kprobe events probe point
seems to be out-of-range. But that is not the correct place to check for it,
it should be done in kprobes afterwards.

kprobe-events allow users to define a probe point on "currently unloaded
module" so that it can trace the event during module load. In this case, the
user will put a probe on a symbol which is not in kallsyms yet and it hits
the within_notrace_func().  As a result, kprobe-events always refuses if
user defines a probe on a "currenly unloaded module".

Fixes: commit 45408c4f9250 ("tracing: kprobes: Prohibit probing on notrace function")
Link: http://lkml.kernel.org/r/153319624799.29007.13604430345640129926.stgit@devbox

	Signed-off-by: Masami Hiramatsu <mhiramat@kernel.org>
	Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>
(cherry picked from commit 6bc6c77cfc6747292d5865d124cc31333b1b8536)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>