LSM: general protection fault in legacy_parse_param
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-496.el8
commit-author Casey Schaufler <casey@schaufler-ca.com>
commit ecff30575b5ad0eda149aadad247b7f75411fd47

The usual LSM hook "bail on fail" scheme doesn't work for cases where
a security module may return an error code indicating that it does not
recognize an input.  In this particular case Smack sees a mount option
that it recognizes, and returns 0. A call to a BPF hook follows, which
returns -ENOPARAM, which confuses the caller because Smack has processed
its data.

The SELinux hook incorrectly returns 1 on success. There was a time
when this was correct, however the current expectation is that it
return 0 on success. This is repaired.

	Reported-by: syzbot+d1e3b1d92d25abf97943@syzkaller.appspotmail.com
	Signed-off-by: Casey Schaufler <casey@schaufler-ca.com>
	Acked-by: James Morris <jamorris@linux.microsoft.com>
	Signed-off-by: Paul Moore <paul@paul-moore.com>
(cherry picked from commit ecff30575b5ad0eda149aadad247b7f75411fd47)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>