soundwire: cadence: Don't overwrite msg->buf during write commands
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-499.el8
commit-author Richard Fitzgerald <rf@opensource.cirrus.com>
commit ba05b39d265bdd16913f7684600d9d41e2796745

The buf passed in struct sdw_msg must only be written for a READ,
in that case the RDATA part of the response is the data value of the
register.

For a write command there is no RDATA, and buf should be assumed to
be const and unmodifable. The original caller should not expect its data
buffer to be corrupted by an sdw_nwrite().

	Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>
	Reviewed-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Link: https://lore.kernel.org/r/20220916103505.1562210-1-rf@opensource.cirrus.com
	Signed-off-by: Vinod Koul <vkoul@kernel.org>
(cherry picked from commit ba05b39d265bdd16913f7684600d9d41e2796745)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>