ASoC: simple-card: Fix up checks for HW param fixups
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-499.el8
commit-author Aidan MacDonald <aidanmacdonald.0x0@gmail.com>
commit 32def55d237e8507d4eb8442628fc2e59a899ea0

The "convert-xxx" properties only have an effect for DPCM DAI links.
A DAI link is only created as DPCM if the device tree requires it;
part of this involves checking for the use of "convert-xxx" properties.

When the convert-sample-format property was added, the checks got out
of sync. A DAI link that specified only convert-sample-format but did
not pass any of the other DPCM checks would not go into DPCM mode and
the convert-sample-format property would be silently ignored.

Fix this by adding a function to do the "convert-xxx" property checks,
instead of open-coding it in simple-card and audio-graph-card. And add
"convert-sample-format" to the check function so that DAI links using
it will be initialized correctly.

Fixes: 047a05366f4b ("ASoC: simple-card-utils: Fixup DAI sample format")
	Acked-by: Kuninori Morimoto <kuninori.morimoto.gx@renesas.com>
	Signed-off-by: Aidan MacDonald <aidanmacdonald.0x0@gmail.com>
	Acked-by: Sameer Pujar <spujar@nvidia.com>
Link: https://lore.kernel.org/r/20221019012302.633830-1-aidanmacdonald.0x0@gmail.com
	Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 32def55d237e8507d4eb8442628fc2e59a899ea0)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>