ASoC: SOF: ipc4: Convert the firmware handling (loader) to library convention
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-499.el8
commit-author Peter Ujfalusi <peter.ujfalusi@linux.intel.com>
commit 5a932cfce4401491c942ddcb7fd3ca669e507b4d

With IPC4 each DSP loadable binary is a library, which contains
ext_manifest section and loadable modules.
The basefw is no exception, it is always library 0 and it can contain
several modules, depending on the firmware build.

The current code assumes only one binary, which is the basefw and has no
concept of libraries.
This patch introduces the library+modules abstraction and represents the
basefw as library for the IPC4 loader codebase.
The basefw loading and handling is not changing, it is still done by the
generic code, but it's information is cloned under the library
representation.

The libraries are managed via XArray to offload the list and ID management.

	Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>
	Reviewed-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>
	Reviewed-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
	Reviewed-by: Chao Song <chao.song@intel.com>
	Reviewed-by: Kai Vehmanen <kai.vehmanen@linux.intel.com>
Link: https://lore.kernel.org/r/20221020121238.18339-10-peter.ujfalusi@linux.intel.com
	Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 5a932cfce4401491c942ddcb7fd3ca669e507b4d)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>