ASoC: tas5805m: rework to avoid scheduling while atomic.
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-499.el8
commit-author Daniel Beer <daniel.beer@igorinstitute.com>
commit 147323792693bf013f60dca160be1d32bd4d180a

There's some setup we need to do in order to get the DSP initialized,
and this can't be done until a bit-clock is ready. In an earlier version
of this driver, this work was done in a DAPM callback.

The DAPM callback doesn't guarantee that the bit-clock is running, so
the work was moved instead to the trigger callback. Unfortunately this
callback runs in atomic context, and the setup code needs to do I2C
transactions.

Here we use a work_struct to kick off the setup in a thread instead.

Fixes: ec45268467f4 ("ASoC: add support for TAS5805M digital amplifier")
	Signed-off-by: Daniel Beer <daniel.beer@igorinstitute.com>
Link: https://lore.kernel.org/r/85d8ba405cb009a7a3249b556dc8f3bdb1754fdf.1675497326.git.daniel.beer@igorinstitute.com
	Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 147323792693bf013f60dca160be1d32bd4d180a)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>