ASoC: SOF: mediatek: Support mt8188 platform
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-499.el8
commit-author Tinghan Shen <tinghan.shen@mediatek.com>
commit 6b43538f0698695fba9aa0c0b29a80c555cf1b63

Add support of SOF on MediaTek MT8188 SoC.
MT8188 ADSP integrates with a single core Cadence HiFi-5 DSP.
The IPC communication between AP and DSP is based on shared DRAM and
mailbox interrupt.

The change in the mt8186.h is compatible on both mt8186 and
mt8188. The register controls booting the DSP core with the
default address or the user specified address. Both mt8186
and mt8188 should boot with the user specified boot in the driver.
The usage of the register is the same on both SoC, but the
control bit is different on mt8186 and mt8188, which is bit 1 on mt8186
and bit 0 on mt8188. Configure the redundant bit has noside effect
on both SoCs.

	Signed-off-by: Tinghan Shen <tinghan.shen@mediatek.com>
	Reviewed-by: Péter Ujfalusi <peter.ujfalusi@linux.intel.com>
	Reviewed-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
	Reviewed-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>
	Reviewed-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>
Link: https://lore.kernel.org/r/20230110084312.12953-3-tinghan.shen@mediatek.com
	Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 6b43538f0698695fba9aa0c0b29a80c555cf1b63)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>