ASoC: SOF: sof-audio: prepare_widgets: Check swidget for NULL on sink failure
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-499.el8
commit-author Peter Ujfalusi <peter.ujfalusi@linux.intel.com>
commit fb4293600cc651cfe4d48ec489f1d175adf6e2f8

If the swidget is NULL we skip the preparing of the widget and jump to
handle the sink path of the widget.
If the prepare fails in this case we would undo the prepare but the swidget
is NULL (we skipped the prepare for the widget).

To avoid NULL pointer dereference in this case we must check swidget
against NULL pointer once again.

Fixes: 0ad84b11f2f8 ("ASoC: SOF: sof-audio: skip prepare/unprepare if swidget is NULL")
	Reported-by: kernel test robot <lkp@intel.com>
	Reported-by: Dan Carpenter <error27@gmail.com>
	Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>
	Reviewed-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>
Link: https://lore.kernel.org/r/20230120102125.30653-1-peter.ujfalusi@linux.intel.com
	Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit fb4293600cc651cfe4d48ec489f1d175adf6e2f8)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>