ASoC: SOF: Protect swidget->use_count with mutex for kcontrol access race
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-499.el8
commit-author Peter Ujfalusi <peter.ujfalusi@linux.intel.com>
commit f94f3915274d22d5cd8b253e0533822b934f5f41

The use_count of the swidget is protect by ALSA core PCM locking with the
exception when an associated kcontrol is changed.

It has been observed that a rightly timed kcontrol access during stream
stop can result of an attempt to send a control update to a widget which
has been freed up between the check of the use_count and the message
sending.

We need to protect the entire sof_widget_setup() and sof_widget_free()
execution to make it safe to rely on the use_count.
Move the code under an _unlocked() function and use a mutex to protect
the execution of the functions for concurrency.
On the control path we need to use the lock only for the kcontrol access,
the widget_kcontrol_setup() op is called with the lock already held.

	Reported-by: Guennadi Liakhovetski <guennadi.liakhovetski@linux.intel.com>
	Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>
	Reviewed-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
	Reviewed-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>
Link: https://lore.kernel.org/r/20230127120031.10709-18-peter.ujfalusi@linux.intel.com
	Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit f94f3915274d22d5cd8b253e0533822b934f5f41)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>