ASoC: tas5720: set bit 7 in ANALOG_CTRL_REG for TAS5720A-Q1 during probe
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-499.el8
commit-author Steffen Aschbacher <steffen.aschbacher@stihl.de>
commit 88f748e38b283702a620e635820f1864bf32db0e

Set the reserved bit 7 in the ANALOG_CTRL_REG for the TAS5720A-Q1 device,
when probing.
The datasheet mentions that the bit should be 1 during reset/powerup.

The device did not initialize before setting this value to 1. So, this
could be a quirk of this device. Or it could be a quirk with the board on
which it was tested.
That is why this patch is separate from the patch that adds support for the
TAS5720A-Q1 device.

	Signed-off-by: Steffen Aschbacher <steffen.aschbacher@stihl.de>
	Signed-off-by: Alexandru Ardelean <alex@shruggie.ro>
Link: https://lore.kernel.org/r/20230128082744.41849-3-alex@shruggie.ro
	Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 88f748e38b283702a620e635820f1864bf32db0e)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>