net: thunderbolt: Use bitwise types in the struct thunderbolt_ip_frame_header
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-500.el8
commit-author Andy Shevchenko <andriy.shevchenko@linux.intel.com>
commit a479f9264bddfd0c266615b83f784222eccf9b91

The main usage of the struct thunderbolt_ip_frame_header is to handle
the packets on the media layer. The header is bound to the protocol
in which the byte ordering is crucial. However the data type definition
doesn't use that and sparse is unhappy, for example (17 altogether):

  .../thunderbolt.c:718:23: warning: cast to restricted __le32

  .../thunderbolt.c:966:42: warning: incorrect type in assignment (different base types)
  .../thunderbolt.c:966:42:    expected unsigned int [usertype] frame_count
  .../thunderbolt.c:966:42:    got restricted __le32 [usertype]

Switch to the bitwise types in the struct thunderbolt_ip_frame_header to
reduce this, but not completely solving (9 left), because the same data
type is used for Rx header handled locally (in CPU byte order).

	Signed-off-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
	Acked-by: Mika Westerberg <mika.westerberg@linux.intel.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit a479f9264bddfd0c266615b83f784222eccf9b91)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>