driver core: allow kobj_to_dev() to take a const pointer
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-500.el8
commit-author Greg Kroah-Hartman <gregkh@linuxfoundation.org>
commit 7ebe49b76a001b10b007193b1771f33e6cbc4f3f

If a const * to a kobject is passed to kobj_to_dev(), we want to return
back a const * to a device as the driver core shouldn't be modifying a
constant structure.  But when dealing with container_of() the pointer
const attribute is cast away, so we need to manually handle this by
determining the type of the pointer passed in to know the type of the
pointer to pass out.

Luckily _Generic can do this type of magic, and as the kernel now
supports C11 it is availble to us to handle this type of build-time type
detection.

	Cc: "Rafael J. Wysocki" <rafael@kernel.org>
	Reviewed-by: Sakari Ailus <sakari.ailus@linux.intel.com>
	Reviewed-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Link: https://lore.kernel.org/r/20221016104126.1259809-1-gregkh@linuxfoundation.org
	Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 7ebe49b76a001b10b007193b1771f33e6cbc4f3f)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>