net/mlx5: Add adjphase function to support hardware-only offset control
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-502.el8
commit-author Rahul Rameshbabu <rrameshbabu@nvidia.com>
commit 8e11a68e2e8abf6510508b75b49ec5953ca262b6

The adjtime function supports using hardware to set the clock offset when
the delta was supported by the hardware. When the delta is not supported by
the hardware, the driver handles adjusting the clock. The newly-introduced
adjphase function is similar to the adjtime function, except it guarantees
that a provided clock offset will be used directly by the hardware to
adjust the PTP clock. When the range is not acceptable by the hardware, an
error is returned.

	Signed-off-by: Rahul Rameshbabu <rrameshbabu@nvidia.com>
	Reviewed-by: Gal Pressman <gal@nvidia.com>
	Reviewed-by: Tariq Toukan <tariqt@nvidia.com>
	Signed-off-by: Saeed Mahameed <saeedm@nvidia.com>
(cherry picked from commit 8e11a68e2e8abf6510508b75b49ec5953ca262b6)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>