net/mlx5: E-switch, Fix wrong usage of source port rewrite in split rules
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-502.el8
commit-author Maor Dickman <maord@nvidia.com>
commit 1313d78ac0c1cfcff7bdece8da54b080e71487c4

In few cases, rules with mirror use case are split to two FTEs, one which
do the mirror action and forward to second FTE which do the rest of the rule
actions and the second redirect action.
In case of mirror rules which do split and forward to ovs internal port or
VF stack devices, source port rewrite should be used in the second FTE but
it is wrongly also set in the first FTE which break the offload.

Fix this issue by removing the wrong check if source port rewrite is needed to
be used on the first FTE of the split and instead return EOPNOTSUPP which will
block offload of rules which mirror to ovs internal port or VF stack devices
which isn't supported.

Fixes: 10742efc20a4 ("net/mlx5e: VF tunnel TX traffic offloading")
Fixes: a508728a4c8b ("net/mlx5e: VF tunnel RX traffic offloading")
	Signed-off-by: Maor Dickman <maord@nvidia.com>
	Reviewed-by: Roi Dayan <roid@nvidia.com>
	Signed-off-by: Saeed Mahameed <saeedm@nvidia.com>
(cherry picked from commit 1313d78ac0c1cfcff7bdece8da54b080e71487c4)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>