arm64: Improve diagnostics when trapping BRK with FAULT_BRK_IMM
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-502.el8
commit-author Will Deacon <will@kernel.org>
commit 0fdb64c2a303e4d2562245501920241c0e507951
Empty-Commit: Cherry-Pick Conflicts during history rebuild.
Will be included in final tarball splat. Ref for failed cherry-pick at:
ciq/ciq_backports/kernel-4.18.0-502.el8/0fdb64c2.failed

When generating instructions at runtime, for example due to kernel text
patching or the BPF JIT, we can emit a trapping BRK instruction if we
are asked to encode an invalid instruction such as an out-of-range]
branch. This is indicative of a bug in the caller, and will result in a
crash on executing the generated code. Unfortunately, the message from
the crash is really unhelpful, and mumbles something about ptrace:

  | Unexpected kernel BRK exception at EL1
  | Internal error: ptrace BRK handler: f2000100 [#1] SMP

We can do better than this. Install a break handler for FAULT_BRK_IMM,
which is the immediate used to encode the "I've been asked to generate
an invalid instruction" error, and triage the faulting PC to determine
whether or not the failure occurred in the BPF JIT.

Link: https://lore.kernel.org/r/20200915141707.GB26439@willie-the-truck
	Reported-by: Ilias Apalodimas <ilias.apalodimas@linaro.org>
	Signed-off-by: Will Deacon <will@kernel.org>
(cherry picked from commit 0fdb64c2a303e4d2562245501920241c0e507951)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>

# Conflicts:
#	arch/arm64/include/asm/extable.h
#	arch/arm64/mm/extable.c